==> 04_Analyses/04_Comparative_Study.Rmd <==
---
title: "04_Comparative_Study"
author: "Dupret Florent"
date: "`r Sys.Date()`"
output:
  html_document: 
    df_print: tibble
    theme: yeti
    highlight: tango
    toc: yes
    toc_depth: 4
    toc_float: yes
    number_sections: yes
    fig_caption: yes
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.show='hold', cache = FALSE, results = 'hold', warning = F, error = F, message = F)
#Cleaning rstudio
rm(list = ls())
#Libraries required
require(ggplot2)
require(tidyverse)
require(FactoMineR)
require(factoextra)
require(visdat)
require(patchwork)
require(SummarizedExperiment)
require(MultiAssayExperiment)
require(pander)
library(magrittr)
library(ggrepel)
```
```{r, include = FALSE}
#Ggplot theme
theme_set(theme_minimal())
#Colors for Cohort and Batch
CohCol <- c("#F8766D", "#00BFC4") |> set_names(c("Stanford", "UMD")) #Ggplot default binary colors
```

This document was simply used to draw the comparative study's colormap.

## Import

```{r}
compara <- readRDS(file = "../03_Data/Compara_assoc.RData")
compara <- compara[[2]]
```

## Plot

```{r}
compara %>%
  mutate(keep = (rowSums(is.na(compara[3:ncol(compara)])) != 10)) |>
  filter(keep == TRUE) |>
  pivot_longer(cols = 3:ncol(compara),
               names_to = "Subcommunity", values_to = "value") |>
  mutate(value = ifelse(is.na(value), 3, value)) |>
  mutate(value = as.factor(value)) |>
  ggplot()+
  aes(x = Subcommunity, y = Metabolite, fill = value)+
  geom_tile(color = "white")+
  scale_x_discrete(position = "top")+
  facet_wrap(facets = vars(approach), ncol = 1, switch = "y")+
  scale_fill_manual(values = c("#F8766D", "purple", "#00BFC4", "white"),
                    labels = c("Stanford", "Both", "UMD", " "),
                    #na.value = "white",
                    aesthetics = c("colour", "fill"))+
  theme(panel.grid.major = element_blank(),
        aspect.ratio = 0.4,
        axis.text.x = element_text(size = 10),
        )+
  labs(fill = "Association \n found in")
```
